ft(done-testing-all) : I just finished testing both trip and booking endpoint APIs with Postman for create, update, delete, and retrieve options [finishes #167825075]

+++ server/controllers/tripController.js
@@ -32,7 +32,7 @@ class tripController {
     if (!newTrip) return response.response(
       res, 401, 'Operation aborted-Invalid inputs', true);
 
-    const id = unique.uniqueID(trips), created_on = fDate(Date());
+    const id = unique.uniqueID(trips), created_on = fDate.curDate();
     newTrip.id = id, newTrip.created_on = created_on;
     newTrip.seats = (()=>{
       let seats = [];
@@ -44,10 +44,7 @@ class tripController {
   }
 
   // update trip info
-  static updateTrip(req, res) {}
-
-  // delete single trip
-  static deleteUnique({is_admin}, req, res, next) {
+  static updateTrip({is_admin}, req, res, next) {
     if (!is_admin) return response.response(
       res, 403, 'Operation aborted-Not allowed', true);
 
@@ -58,6 +55,50 @@ class tripController {
     const trip_update = tripSchema.update(req.body);
     if (!trip_update) return response.response(
       res, 401, 'Operation aborted-Invalid inputs', true);
+
+    const trip = filter.filterTrips(is_admin, trip_id);
+    if (!trip) return response.response(
+      res, 404, 'Operation aborted-Resource not found', true);
+    trip_id = null;
+
+    if (sc = trip_update.seating_capacity) {
+      if (sc === trip.seating_capacity) return response.response(
+        res, 400, 'Resource unchanged, already up-to-date', true);
+
+      if (sc > trip.seating_capacity) {
+        (() => {
+          for (let i=(trip.seating_capacity+1); i<=sc; i++) 
+            trip.seats.push(i);
+        })();
+        trip.seating_capacity = sc;
+      }
+      else {
+        if ((trip.seating_capacity - trip.seats.length) > sc) {
+          return response.response(
+            res, 400, "Resource's changes not appliable", true);
+        }
+        if ((trip.seating_capacity - trip.seats.length) === sc) {
+          return response.response(
+            res, 400, 'Resource unchanged, already up-to-date', true);
+        }
+        filter.updateTripSeatings(trip, sc);
+      }
+    }
+
+    if (bln = trip_update.bus_licence_number) {
+      trip.bus_licence_number = bln;
+    }
+    return response.response(res, 200, trip);
+  }
+
+  // delete single trip
+  static deleteUnique({is_admin}, req, res, next) {
+    if (!is_admin) return response.response(
+      res, 403, 'Operation aborted-Not allowed', true);
+
+    const trip_id = isParam.intParam(req.params);
+    if (!trip_id) return response.response(
+      res, 400, 'Operation aborted-Bad parameter', true);
 
     const trip = filter.filterTrips(is_admin, trip_id);
     if (!trip) return response.response(
@@ -75,7 +116,23 @@ class tripController {
   }
 
   //delete all
-  static deleteAll(req, res) {}
+  static deleteAll({is_admin}, req, res, next) {
+    if (!is_admin) return response.response(
+      res, 403, 'Operation aborted-Not allowed', true);
+
+    trips.forEach(trip => {
+      if (trip.status === 'cancelled') return;
+      trip.status = 'cancelled';
+
+      bookings.forEach(book => {
+        if ((book.status === 'cancelled') || (
+          book.trip_id !== trip.id)) return;
+        
+        book.status = trip.status;
+      });
+    });
+    return response.response(res, 200, trips);
+  }
 }
 
 export default tripController;
